refactor: split code of exports_info
