+++ crates/rspack_core/src/exports/mod.rs
@@ -1,0 +1,7 @@
+mod export_info;
+mod exports_info;
+mod utils;
+
+pub use export_info::*;
+pub use exports_info::*;
+pub use utils::*;
